+++ l-app/language/english/general/auth_lang.php
@@ -3,17 +3,14 @@
 $lang['signin_title'] = 'Sign In to Administrator';
 $lang['signin_welcome'] = 'Welcome back! Please sign in to continue.';
 $lang['have_an_account'] = 'Don\'t have an account? <a href="#" class="tx-dark tx-semibold">Create an Account</a>';
-
 $lang['forgot_title'] = 'Forgot Password';
 $lang['forgot_content'] = 'Enter your verified email address and we will send your account information.';
 $lang['forgot_password'] = 'Forgot password';
 $lang['enter_your_email'] = 'Enter your email address';
 $lang['back_to_login'] = 'Back to sign in';
-
 $lang['username'] = 'Username';
 $lang['password'] = 'Password';
 $lang['login_email'] = 'Email';
-
 $lang['message_login_error'] = 'The password you entered is incorrect';
 $lang['message_forgot_error'] = 'Email not found';
 $lang['message_forgot_success'] = 'Password information has been sent to your email';