+++ l-app/language/english/general/ui_lang.php
@@ -8,7 +8,6 @@ $lang['ui_users'] = 'Users';
 $lang['ui_module'] = 'Module';
 $lang['ui_apperance'] = 'Apperance';
 $lang['ui_component'] = 'Component';
-
 $lang['ui_all'] = 'All';
 $lang['ui_check_all'] = 'Checked All';
 $lang['ui_select_all'] = 'Select All';
@@ -17,7 +16,6 @@ $lang['ui_add'] = 'Add';
 $lang['ui_edit'] = 'Edit';
 $lang['ui_update'] = 'Update';
 $lang['ui_action'] = 'Action';
-
 $lang['ui_active'] = 'Active';
 $lang['ui_deactive'] = 'Deactive';
 $lang['ui_yes'] = 'Yes';
@@ -25,17 +23,12 @@ $lang['ui_no'] = 'No';
 $lang['ui_publish'] = 'Publish';
 $lang['ui_draft'] = 'Draft';
 $lang['ui_banned'] = 'Banned';
-
 $lang['ui_male'] = 'Male';
 $lang['ui_female'] = 'Female';
-
 $lang['confirmation_deletion'] = 'Confirmation Deletion';
 $lang['confirmation_deletion_content'] = 'Are you sure want to delete this data ?';
-
 $lang['form_message_add_success'] = 'Data has been successfully added';
 $lang['form_message_update_success'] = 'Data has been successfully updated';
 $lang['form_message_delete_success'] = 'Data has been successfully deleted';
-
 $lang['form_message_submit_error'] = 'Please enter data correctly';
-
 $lang['ui_environment_development_info'] = '<h4 class="tx-medium tx-uppercase tx-dark mb-2"><i class="cificon licon-alert-triangle mr-2 text-danger"></i>Attention !</h4><div>This website is still in <code>DEVELOPMENT</code> mode.<br>To go online, please change the <code>ENVIRONMENT</code> mode to <code>PRODUCTION</code>.</div>';