+++ l-content/plugins/filemanager/config/config.php
@@ -29,7 +29,7 @@ elseif ($_SESSION['filemanager_access']['read_access'] == TRUE)
 
 	mb_internal_encoding('UTF-8');
 	mb_http_output('UTF-8');
-	mb_http_input('UTF-8');
+	// mb_http_input('UTF-8');
 	mb_language('uni');
 	mb_regex_encoding('UTF-8');
 	ob_start('mb_output_handler');